+++ src/Makefile.am
@@ -156,7 +156,7 @@ BITCOIN_CORE_H = \
   ui_interface.h \
   uint256.h \
   undo.h \
-  unlimited.h \
+  xthinblocks.h \
   util.h \
   utilmoneystr.h \
   utilstrencodings.h \
@@ -215,7 +215,7 @@ libbitcoin_server_a_SOURCES = \
   torcontrol.cpp \
   txdb.cpp \
   txmempool.cpp \
-  unlimited.cpp \
+  xthinblocks.cpp \
   validationinterface.cpp \
   versionbits.cpp \
   $(BITCOIN_CORE_H)
