+++ src/net.cpp
@@ -19,7 +19,7 @@
 #include "scheduler.h"
 #include "ui_interface.h"
 #include "utilstrencodings.h"
-#include "unlimited.h"
+#include "xthinblocks.h"
 
 #ifdef WIN32
 #include <string.h>
