+++ src/txmempool.cpp
@@ -15,7 +15,7 @@
 #include "util.h"
 #include "utilmoneystr.h"
 #include "utiltime.h"
-#include "unlimited.h"
+#include "xthinblocks.h"
 #include "version.h"
 
 using namespace std;
